Java bootclasspath: use NestedSet
Before this change, we start with a list, then convert it to a nested set, and then convert it back to a list. This causes unnecessary copying as well as lock contention (because NestedSetBuilder.wrap tries to intern the intermediate immutable lists).

Unfortunately, the Starlark API exposes the bootclasspath as a list, so round-trips through Starlark still have to convert back and forth to/from nested set.

PiperOrigin-RevId: 283775541