Report on download cache look up and archive extraction
Checking an entry in the download cache or a distdir might take a
while, as the hash of that file has to be computed. Extracting an
archive can take even more time, as that usually implies a lot of
disk IO. Inform the user about the progress here to avoid confusing
this delay with network activity.

Change-Id: I86316529309960a8a501b96ae752a12a554197b9
PiperOrigin-RevId: 284140963