bazel C++: remove redundant tool methods like ar_executable()
The CcToolchainProvider defines each of these attributes both as a
field and as a method, so x.f is equivalent to x.f(), which breaks an
invariant of the Starlark spec that x.f() should be decomposable into
y=x.f; y().

This change removes the methods.

RELNOTES[INC]:
The following attributes of CcToolchainProvider, which formerly were
accessible both as fields (x.f) and as methods (x.f()), are now only fields:

            ld_executable
       objcopy_executable
      compiler_executable
  preprocessor_executable
            nm_executable
       objdump_executable
            ar_executable
         strip_executable

Remove the call parens (x.f() -> x.f) if your code uses them as methods.

PiperOrigin-RevId: 284776115