bazel syntax: eliminate two subclasses of BuiltinFunction
This change replaces two test-only subclasses of BuiltinFunction with
StarlarkCallables. (The remaining two subclasses, rule and
repository_rule, will be dealt with in a follow-up.)

Also, change type of rule's and repository_rule's "implementation" parameter
to StarlarkFunction, and eliminate casts that already made this assumption.

Also, give StarlarkCallable.getLocation a default value of BUILTIN.

PiperOrigin-RevId: 285022081