bazel: stop implementing BaseFunction.call
Most subclasses of BaseFunction have a signature of f(**kwargs), for
which BaseFunction's "inner" protocol is overkill. This change causes
them to implement StarlarkCallable.callImpl directly.

For now, they still override BaseFunction, because it defines other
methods (e.g. repr), and because some places in the code check for
"instanceof BaseFunction" when they should really be more general
(StarlarkCallable) or more specific (StarlarkFunction).

PiperOrigin-RevId: 285259322