Log (but don't throw) when encountering an unknown attribute on a merged styleable
Styleables are merged, which can lead to situation where a styleable refers to
an attribute from a transitive dependency.  Rather than add a dep for unused
attributes to make RPBB happy, we'll just drop it and let javac complain if
they're actually used.  Styleables for custom views shouldn't be merged
anyway, and won't be when they're marked private.

PiperOrigin-RevId: 285885337