Let shellzelisk fallback to bazel-real if it's the requested version.
This means that if you "apt-get install bazel" and that happens to be version "2.0.0" and in your `.bazelversion` file you have "2.0.0", then the new Bazel wrapper will just use the Bazel binary installed as the "latest" Bazel by "apt-get install bazel" and not ask the user to install the same version of Bazel as a side-by-side binary using "apt-get install bazel-2.0.0".

The same applies when using the Bazel installer or any other setup that uses the wrapper script.

Closes #10356.

RELNOTES: None.
PiperOrigin-RevId: 285941301