Remove CC_METADATA_COLLECTOR from sh_binary
This was introduced in unknown commit as part of a more complex change,
and I believe that this part was unnecessary.

The InstrumentedFilesCollector performs two functions: it combines
coverage information from direct deps, and it collect local information
from the current rule. However, the local information collector only
supports C++, Java, and Go actions. For rules that do not locally create
any of these actions, the local collector is a no-op.

Since sh_binary does not generate any C++ actions directly, this change
should be a no-op.

PiperOrigin-RevId: 285967632