Undo ArtifactNestedSetFunction's making concurrent map weak-keyed.
The concurrent map being weak-keyed means the keys are compared by reference. This would cause a problem when the same blaze instance is used to build different targets:

e.g.
1. Build //foo. State of ArtifactNestedSetFunction.artifactSkyKeyToValueOrException: { A1: VOE1, A2: VOE2 }, as a result of evaluating NestedSet NS1: [A1, A2]
2. Build //bar, which contains NS1_: [A1_, A2_]. Assume NS1_.equals(N1) => Skyframe recognizes this ArtifactNestedSetKey as a cache hit and won't run ArtifactNestedSetFunction.compute(NS1_'s key), instead simply returning the non-null cached SkyValue.
=> ArtifactNestedSetFunction.artifactSkyKeyToValueOrException remains { A1: VOE1, A2: VOE2 }

Now, we try to access this map, looking for A1_ and A2_ in ActionExecutionFunction, since a non-null ArtifactNestedSetValue means a promise that the entries for children artifacts exist in the map (invariant).

When we try to access the map with A1_ and A2_, it fails since the map goes by reference
=> Exception is thrown.

RELNOTES: None
PiperOrigin-RevId: 286046864