bazel packages: delete AstParseResult
This class held the result of parsing: a syntax tree, a list of scan/parse errors,
and a list of posts. However, the list of posts was empty and the list of errors
was a copy of file.errors(). The whole thing served only to obscure.

Once removed, createPackage became trivial and the sole call was inlined.

Also, rename various AST-related names to "syntax".

PiperOrigin-RevId: 286077843