bazel skylarkinterface processor: check parameter types match annotations
The SkylarkCallable annotation processor historically did not
check statically that the annotation types in Param.type and
Param.allowed_types are assignable to the parameter variable,
resulting in bugs such as
https://github.com/bazelbuild/bazel/issues/9463
in which the annotation type check succeeds but then the
reflective call explodes.

This change causes the annotation processor to ensure
that the annotation types are assignable to the variable,
and fixes up a number of places that violated the rule.
(A prettier fix for the SkylarkAttr violations is in unknown commit.)

Also:
- fold verifyParamGenericTypes into the mass of parameter checks.
- allow reporting of errors against any element.
- report multiple errors in one run; don't throw exceptions.
- report parameter errors against parameters, not the method.
- rename parameter to paramAnnot and methodParam to param.
- use proper type equality not string equality checks.
PiperOrigin-RevId: 286193084