bazel repository_rule: don't introspect on FuncallExpression
This CL changes the rule class name heuristic using during instantiation
of non-exported repository_rules (an unintended feature).

A repository rule ought to be subject to the same "exported"
check applied to an ordinary package rule, that is, one should not
be able to instantiate it unless it has been "exported",
which means assigned to a variable at top level.
However, the necessary check was sadly omitted and now several
projects exploit the weakness to create and instantiate repository
rules back to back without an intervening "export".
(Never mind that the current assignment-based export semantics
are problematic.)

To guess the name of a non-exported rule class, the previous code
hackily introspected the syntax tree of the call expression
f() or x.f(), so in the code below it would infer a rule class name
of "r":

  def f():
    r = repository_rule(impl=_r_impl)
    r(name="name", ...)

Soon, the syntax tree will no longer be available to the callee,
so the heuristic must change. Now, it uses "unexported-_r_impl",
based on the name of the rule implementation function.

Users should not be instantiating non-exported repository rules.
The correct usage, which has not changed, is to declare the
repository rule at top level:

  r = repository_rule(impl=_r_impl)

  def f():
     r(...) # instantiates an "r" repository rule

PiperOrigin-RevId: 286239162