bazel syntax: define and use Starlark.errorf
This change defines an API helper function that returns an
EvalException with no location and an error message produced by
formatting the arguments.

It also updates all applicable places within the package to use it:
when creating an Exception whose location is either null or obviously
unneeded (such as within a built-in), with no "cause" exception.

This eliminates all uses of useLocation=true within the package
except the one in 'print'. We should be able to eliminate useLocation
globally once the location can be retrieved from the callstack
for the few functions (like print) that need it.

Exceptions thrown by the evaluator itself, including operators
like EvalUtils.getIndex, should continue to report the location of the
failed operator. (The operators are not entirely consistent about this.
At some point I will remove the Location parameters to the operators and
make the evaluator responsible for adding locations, if necessary by
catching and rethrowing.)

PiperOrigin-RevId: 286441023