bazel syntax: don't pass call syntax to function calls
It should never have been necessary to provide a syntax tree when
making a function call, and doing so is an obstacle to optimizations
(such as compilation). It also encouraged metaprogramming hacks
such as the "generator_name" attribute (reimplemented as a static
analysis in CL 283349613) and SkylarkRepositoryModule's reinvention of the
already problematic 'exported name' wheel through function introspection
(see CL 283533234).

This change replaces FuncallExpression with Location in the core
calling machinery and in all SkylarkCallable-annotated functions.
Places where the call Location was used only to throw EvalException have
been replaced with null as the interpreter fills these in automatically.

(Later changes will remove even Location from the calling convention.
Instead, the very few remaining clients that need it will be able to
obtain the location of the program counter in the nth enclosing call
frame using the debug API.)

Also:
- remove StarlarkCallable.useAst.
- use Location.BUILTIN for all calls through StarlarkCallbackHelper.

This is a breaking API change for copybara.

BEGIN_PUBLIC
bazel syntax: don't pass call syntax to function calls
END_PUBLIC

PiperOrigin-RevId: 286574619