Make sure to wait for all subprocesses in the process-wrapper.
This is take 2 of https://github.com/bazelbuild/bazel/commit/7828118ea8f4150aa1eae4ab3133935ffe221eb5, which was rolled back because it broke
downstream Android tests.  This change differs from the original one
in that, on Linux, we make sure to kill any subprocess that has escaped
the process group (as otherwise we can get stuck due to the new use of
the child subreaper feature).

When we kill the process group, make sure to wait until the results
of all subprocesses have been collected to ensure all subprocesses are
stopped before the process-wrapper exits.

I have been unable to come up with a test case to show that this does
anything useful though (in particular because the process-wrapper uses
SIGKILL against the processes instead of SIGTERM so I cannot install
any handlers to mess things up). Conceptually, the problem existed, but
it is a race condition that is very hard to trigger. But I'm adding
some unit tests for the internals.

Fixes: https://github.com/bazelbuild/bazel/issues/10245.

RELNOTES: None.
PiperOrigin-RevId: 286596638