Remove dep fingerprinting, part 2 of 2.
Removes SkyValue#getValueFingerprint and Version#getFingerprint. This should also free a little bit of memory that was used in configured target values.

getValueFingerprint is kept in two SkyValue subclasses (FileStateValue and FileArtifactValue) where it is actually used for another purpose. However, the method is removed from the SkyValue interface and other classes where it is not being used.

PiperOrigin-RevId: 286601483