Load all aapt2 *.flat files before forwarding them to consumers
aapt2's protobuf definitions have "Visibility" on "ResourceTable" but not
"CompiledFile" since this is how the input files and final resources.arsc+res/
work.  However, RPBB has a unified model which combines all types of resources
into a single map.  This change will allow us to figure out the correct
visibility for "CompiledFile" resources.

This happened to fix a bug in minSdk level for density by relying on the primary
source (Android framework) instead of com.android.ide.* which is wrong and
unmaintained.  In practice, this meant that resource conflicts for "color"
may have been undetected because they can be defined as plain hex values in a
ResourceTable or as a ColorStateList in a standalone file, and thus two
different resource qualifier processors were used.

PiperOrigin-RevId: 286661439