Add a LostInputCheck so failed rewindable executions can avoid side effects
It delegates to checking an action filesystem for lost inputs, if such a
filesystem is in use. The check is available in ActionExecutionContext
and SpawnExecutionContext.

Also, this changes the return type of getDepOwners from Collection to
ImmutableSet to more clearly embody its properties: it's unordered and
should have no duplicates.

Action execution involves two failure exception base types, ExecException
and ActionExecutionException. Like before, there remain two "lost inputs"
exception variants, one for each. There was, and is, a way from
converting in one direction, and now there is a way of safely converting
in both directions. {Exec,ActionExecution}Exception probably instead
should be collapsed to one type, but that's a much bigger refactoring
that this avoids.

RELNOTES: None.
PiperOrigin-RevId: 287190381