Use "compiled" versions of resources for detecting resource conflicts
The previous code was using the raw "uncompiled" sources which are redundant
and can lead to bypassing some processing steps.  In particular, by falling
back to checking the raw sources, this ignored visibility as a criterion for a
resource conflict.

To save memory, we just store hashes/fingerprints of artifacts.  Since
AndroidCompiledDataDeserializer can be in the critical path (R class generation)
we only read these artifacts when needed.

PiperOrigin-RevId: 287609977