bazel skylarkinterface: enforce structField => !useLocation
This change adds a compile-time check that SkylarkCallable-annotated
"struct fields" cannot request the location of their "call".

That's always been part of the design, and rightly so, because unlike
a method call, which is actively executed within the context of some
thread that has a stack of locations, a field is a passive thing,
a part of a data structure, and it may be accessed by a Java thread
that has nothing to do with Starlark threads or locations.

Unfortunately, as so often is the case, we forgot to enforce this
rigorously, and "nature finds a way":

- repository_ctx.os used the location in recorded events,
  which appear in the workspace log file produced by
  --experimental_workspace_rules_log_file; see
  https://docs.bazel.build/versions/master/workspace-log.html.
  The log now reports the location of rule instantiation.

- google_cc_common.ProtoCcInfo used it to enforce a whitelist.
  This field has been replaced by a method, google_cc_common.get_ProtoCcInfo().
  The field will briefly continue to exist but not enforce the whitelist.
  The sole current use is one file in Google code.

Also, in SkylarkCallable annotation processor:
- move annotation processor docs into annotation API, and revise.
- stop using string comparisons for type queries.
- move checking of special (n?e "extra interpreter") parameters into
  verifyParameters, making a single pass over the parameters, reporting
  multiple errors.
- accept any supertype of Tuple<Object> for *args parameter.
- accept any supertype of Dict<String, Object> for **kwargs parameter.
  Various uses updated to the more refined type.
- report parameter errors against the parameter, not the method.
- make error messages more concise and consistent.

PiperOrigin-RevId: 287668281