Automated rollback of commit 9c1853aa2fe49cf80dd467bb4020eb0822d53881.
*** Reason for rollback ***

b/147116236

*** Original change description ***

Make sure to wait for all subprocesses in the process-wrapper.

This is take 2 of https://github.com/bazelbuild/bazel/commit/7828118ea8f4150aa1eae4ab3133935ffe221eb5, which was rolled back because it broke
downstream Android tests.  This change differs from the original one
in that, on Linux, we make sure to kill any subprocess that has escaped
the process group (as otherwise we can get stuck due to the new use of
the child subreaper feature).

When we kill the process group, make sure to wait until the results
of all subprocesses have been collected to ensure all...

***

PiperOrigin-RevId: 288053699