Shared actions must not emit rewound events if they aren't primary
To be more like other action event types, ActionRewoundEvent should only
be emitted for the action which is primary (i.e. won the race to
evaluate) among its set of shared actions.

To do so, an action evaluation observing lost inputs compares
ActionLookupData values between the lost inputs exception and itself to
determine whether the current action is primary. It emits
ActionRewoundEvent only if it is.

Also, handleLostInputs now only invalidates SkyframeActionExecutor's
buildActionMap if the action is primary, to fix a theoretical bug. If
non-primary actions evict, then they might evict the retried action,
which would allow >1 action in a shared action set to evaluate
simultaneously.

RELNOTES: None.
PiperOrigin-RevId: 288387504