bazel syntax: remove StarlarkThread.update from API
This function served three cases:
1) binding a local. This is now served by updateLexical, which is private.
2) binding a global. Now: thread.getGlobals().put()
3) binding an unresolved identifier. Now: updateUnresolved, which is private.
This reduces exposure of StarlarkThread internals in preparation
for an improved representation of the environment.

It is also a step towards decoupling Module and StarlarkThread:
(Logically, a thread does not have a module, but every function and file does.)
Eventually we will get rid of thread.getGlobals.

Also:
- StarlarkThread.moduleLookup is private. Now: thread.getGlobals.lookup()
- Module.getDirectBindings is renamed to 'lookup'.
  (Module.get looks in the universe too.  Eventually that will go away.)
PiperOrigin-RevId: 288418467