Convert two helper methods in bazel_workspaces_test.sh to Python 3.
The shell script ran some Python 2.x only code using `python`, but the `python` executable is not guaranteed to be Python 2.x. On Arch Linux, it is Python 3.x.

Let's convert the code to Python 3 and then explicitly run it using `python3`, which should work on all systems where any version of Python 3 is installed (which is OK to assume now after the Python 2 deprecation).

RELNOTES: None.
PiperOrigin-RevId: 288448365