SkyframeExecutor: don't create SkyKey list
Remove ConfiguredTargetValue.keys() because it
merely creates an ImmutableList copy of its
argument. The function's only caller did not need
that ImmutableList, because it just passed the
result of .keys() to an ArrayList ctor.

Also, do not create a List of SkyKeys to evaluate
because the other side expects an Iterable and
makes an ImmutableList copy anyway.

PiperOrigin-RevId: 288466171