Add support for a new dependency in generated protos
The proto package is a hodgepodge of stuff.
All of the Internal_{ProtoReflect,ProtoIface,ProtoImpl}_XXX declarations
should never be used by users and is only used by generated .pb.go files.

This is a limited roll-forward of CL/242006724, but introduces protodeps
as an *additional* dependency (rather than a *replacement*) for
generated .pb.go files. CL/242006724 was rolled back because it broke
targets that depended on //net/proto2/go:proto without explicitly declaring
such a dependency. These targets accidentally work since the dependency
happened to be satisfied by a go_proto_library rule.

The removal of the proto package as a dependency will occur in a future CL.

RELNOTES: Generated Go protobufs now depend on //net/proto2/go:protodeps in addition to //net/proto2/go:proto
PiperOrigin-RevId: 288539220