Make cquery 'somepath' more accurate.
Specifically, make somepath(//foo, //bar) report results
when //bar isn't built in the top-level configuration (because
of a transition).

To really work properly, this requires --universe_scope:

$ blaze cquery 'somepath(//foo, //bar)' -- universe_scope=//foo

Otherwise, //bar also gets built in the top-level configuration and
somepath looks for *that* version, which is not the one in //foo's
deps.

RELNOTES: cquery 'somepath' returns more reliable results when the
dep has a different configuration than the parent. To get a result for
`somepath(//foo, //bar`) where //bar isn't in the top-level configuration,
run your query with `--universe_scope=//foo`. See cquery docs for details.
PiperOrigin-RevId: 288560710