enforceLabelAttributes: Evaluate labels in Dict's key set as well
Since we support label_keyed_string_dict as an attribute type of a Starlark rule, we also need to evaluate the labels in advance. This helps avoiding restarting the SkylarkRepositoryFunction, which is potentially expensive (maybe including extracting a compressed file, as we've seen in the bug).

Fixes https://github.com/bazelbuild/bazel/issues/10515

RELNOTES: None
PiperOrigin-RevId: 288674429