Improve printing test results summary.
Print up to 5 first failed-to-build tests.
This will help to locate the problems with tests themselves, while keeping the amount of output small.
(We do not want to print all names of failed-to-build tests in case when some common intermediate target is a problem.)

Currently, the user has to scroll a verbose output to locate the failed-to-build target.

Closes #10511.

PiperOrigin-RevId: 288677606