SkyframeBuildView: move action conflict checking
Bazel will now delay checking for action conflicts
until it has collected "good" Aspects and CTs and
has created the PackageRoots, because doing so
will let us move the collection and creation logic
into a SkyFunction and so reduce code outside of
Skyframe between Analysis and Execution, and
thereby allow us to maybe merge the two phases
some time in the future.

I believe the collection and creation logic does
not actually depend on the result of action
conflict checking, making this change safe.

I believe this because the collection logic does
not actually use the 'badActions' map, and while
it looks up SkyValues from the EvaluationResult,
those are supposed to be immutable, so the
conflict checker could not be storing anything in
them, therefore the collection logic cannot
possibly be filtering CTs with bad actions.

Therefore "good" in the variable names "goodCts"
and "goodAspects" is misleading, and also the
comment that claimed the "goodCts" collection
logic was filtering out CTs with bad actions was
wrong, so I renamed these variables and I removed
the comment.

I also fixed a minor bug: the 'aspects' list
(previously: 'goodAspects') was initialized with
an expected size of 'ctKeys.size()', now it is
initialized with the 'aspectKeys.size()'.

PiperOrigin-RevId: 288687661