bazel syntax: remove Location from Starlark.call
This change removes the Location parameter from the call operator, Starlark.call.
The thread's call stack now records the position of the program
counter of each active frame. Clients may access it through
StarlarkThread.getCallerLocation. The frame location is updated only
at certain key points, such as at function entry, breakpoints,
and when making a call.

The stack of call frames in StarlarkThread is no longer off-by-one.
That is, we now push a frame for the implicit function (called "<toplevel>")
representing the top-level statements of a file, and for the implicit
function (called "<expr>") representing an isolated expression.
In each case we materialize a StarlarkFunction and call it in the usual way.
All Starlark evaluation now occurs in the context of a function call
with a call frame, bringing us a step closer to the "exec API" documented
in comments in Starlark.java.

Unfortunately, because identifiers are not uniformly resolved prior
to execution, we must treat <toplevel> functions specially:
they do not get a LexicalFrame, and assignments to unresolved
identifiers are assumed to refer to globals.
This internal hack will be removed in due course.

For now, the StarlarkCallable.{fast,}call methods still provide the
redundant Location parameter to callees, but this will be removed in a
follow-up change.

Also:
- remove Location parameter from SkylarkClassObject.getValue.
- Eval: unify top-level + non-toplevel execStatements logic and simplify.
- rewrite and augment one debugger test, which was too low-level.
- avoid useLocation=true throughout lib.syntax.  Use useStarlarkThread instead.

BEGIN_PUBLIC
bazel syntax: remove Location from Starlark.call
END_PUBLIC

PiperOrigin-RevId: 288710067